move receiver/scrapererror to scraper/scrapererror
