+++ scraper/scrapererror/doc.go
@@ -2,4 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Package scrapererror provides custom error types for scrapers.
-package scrapererror // import "go.opentelemetry.io/collector/receiver/scrapererror"
+package scrapererror // import "go.opentelemetry.io/collector/scraper/scrapererror"
