+++ scraper/scrapererror/partialscrapeerror_test.go
