+++ scraper/scrapererror/scrapeerror_test.go
